refactor(cluster, workflow): simplify backport internals and test fixtures
