+++ packages/cluster/test/ClusterWorkflowEngine.test.ts
@@ -12,7 +12,6 @@ import {
 import { Rpc } from "@effect/rpc"
 import { assert, describe, expect, it } from "@effect/vitest"
 import { Activity, DurableClock, DurableDeferred, Workflow } from "@effect/workflow"
-import * as WorkflowEngineModule from "@effect/workflow/WorkflowEngine"
 import { WorkflowEngine, WorkflowInstance } from "@effect/workflow/WorkflowEngine"
 import {
   Context,
@@ -32,12 +31,14 @@ import {
 } from "effect"
 import * as Cause from "effect/Cause"
 import * as Duration from "effect/Duration"
+import * as WorkflowEngineContractTest from "../../workflow/test/WorkflowEngineContractTest.js"
 import * as Abandon from "../src/internal/clusterAbandon.js"
 import * as RunnerHealth from "../src/RunnerHealth.js"
 import * as RunnerStorage from "../src/RunnerStorage.js"
 import { abandonmentCause, MemoryLive } from "./fixtures/abandonment.js"
 import { makeRequest } from "./fixtures/message-storage.js"
 import { runFixture } from "./fixtures/run-fixture.js"
+import { makeEngine, makeMemoryEngine } from "./fixtures/workflow-engine.js"
 
 describe.concurrent("ClusterWorkflowEngine", () => {
   it.effect("should run a workflow", () =>
@@ -459,7 +460,7 @@ describe.concurrent("ClusterWorkflowEngine", () => {
     }).pipe(Effect.provide(TestWorkflowLayer)))
 })
 
-const TestShardingConfig = ShardingConfig.layer({
+const TestWorkflowEngine = makeMemoryEngine({
   shardsPerGroup: 300,
   availableShardGroups: ["default", "workflow"],
   assignedShardGroups: ["default", "workflow"],
@@ -468,15 +469,6 @@ const TestShardingConfig = ShardingConfig.layer({
   entityMessagePollInterval: 5000,
   sendRetryInterval: 100
 })
-
-const TestWorkflowEngine = ClusterWorkflowEngine.layer.pipe(
-  Layer.provideMerge(Sharding.layer),
-  Layer.provide(Runners.layerNoop),
-  Layer.provideMerge(MessageStorage.layerMemory),
-  Layer.provide(RunnerStorage.layerMemory),
-  Layer.provide(RunnerHealth.layerNoop),
-  Layer.provide(TestShardingConfig)
-)
 
 class SendEmailError extends Schema.TaggedError<SendEmailError>("SendEmailError")("SendEmailError", {
   message: Schema.String
@@ -870,20 +862,14 @@ const TestWorkflowLayer = EmailWorkflowLayer.pipe(
 )
 
 describe("abandonment", () => {
-  const EngineLive = ClusterWorkflowEngine.layer.pipe(
-    Layer.provideMerge(Sharding.layer),
-    Layer.provide(RunnerStorage.layerMemory),
-    Layer.provide(RunnerHealth.layerNoop),
-    Layer.provide(Runners.layerNoop),
-    Layer.provide(ShardingConfig.layer({
-      shardsPerGroup: 1,
-      entityTerminationTimeout: 0,
-      entityMessagePollInterval: 100,
-      entityReplyPollInterval: 100,
-      refreshAssignmentsInterval: 100,
-      sendRetryInterval: 10
-    }))
-  )
+  const EngineLive = makeEngine({
+    shardsPerGroup: 1,
+    entityTerminationTimeout: 0,
+    entityMessagePollInterval: 100,
+    entityReplyPollInterval: 100,
+    refreshAssignmentsInterval: 100,
+    sendRetryInterval: 10
+  })
 
   for (const recovery of ["catchAllCause", "exit"] as const) {
     for (const masked of [false, true]) {
@@ -1338,326 +1324,17 @@ describe("workflow send-time abandonment", () => {
   }
 })
 
-describe("workflow engine contract", () => {
-  const clusterEngine = ClusterWorkflowEngine.layer.pipe(
-    Layer.provideMerge(Sharding.layer),
-    Layer.provide(Runners.layerNoop),
-    Layer.provideMerge(MessageStorage.layerMemory),
-    Layer.provide(RunnerStorage.layerMemory),
-    Layer.provide(RunnerHealth.layerNoop),
-    Layer.provide(ShardingConfig.layer({
-      shardsPerGroup: 10,
-      entityTerminationTimeout: 0,
-      entityMessagePollInterval: 5000,
-      entityReplyPollInterval: 10,
-      sendRetryInterval: 10
-    }))
-  )
-  const tick = Effect.flatMap(
+WorkflowEngineContractTest.suite({
+  name: "cluster",
+  engineLayer: makeMemoryEngine({
+    shardsPerGroup: 10,
+    entityTerminationTimeout: 0,
+    entityMessagePollInterval: 5000,
+    entityReplyPollInterval: 10,
+    sendRetryInterval: 10
+  }),
+  tick: Effect.flatMap(
     Effect.serviceOption(Sharding.Sharding),
     (sharding) => Option.isSome(sharding) ? sharding.value.pollStorage : Effect.yieldNow()
   )
-  const settle = Effect.gen(function*() {
-    yield* tick
-    yield* TestClock.adjust(1)
-  })
-
-  const backend = "cluster"
-  const engineLayer = clusterEngine
-  const awaitResult = <A, E>(
-    workflow: {
-      readonly poll: (
-        id: string
-      ) => Effect.Effect<Workflow.Result<A, E> | undefined, never, WorkflowEngineModule.WorkflowEngine>
-    },
-    executionId: string,
-    tag: "Suspended" | "Complete",
-    ready: () => boolean = () => true
-  ) =>
-    Effect.gen(function*() {
-      let result = yield* workflow.poll(executionId)
-      for (let i = 0; i < 100 && (result?._tag !== tag || !ready()); i++) {
-        yield* settle
-        result = yield* workflow.poll(executionId)
-      }
-      assert.strictEqual(result?._tag, tag)
-      assert.isTrue(ready())
-      return result
-    })
-
-  describe(`${backend} suspension and deferred completion`, () => {
-    for (
-      const { childCount, concurrency, waves } of [
-        { childCount: 3, concurrency: "unbounded" as const, waves: [3] },
-        { childCount: 5, concurrency: 2, waves: [2, 4, 5] }
-      ]
-    ) {
-      it.effect(`activity fan-out suspends and replays with concurrency ${concurrency}`, () =>
-        Effect.gen(function*() {
-          const Parent = Workflow.make({
-            name: "BackportFanOutParent",
-            payload: {},
-            success: Schema.Array(Schema.Number),
-            idempotencyKey: () => "parent"
-          })
-          const Child = Workflow.make({
-            name: "BackportFanOutChild",
-            payload: { index: Schema.Number },
-            success: Schema.Number,
-            idempotencyKey: ({ index }) => String(index)
-          })
-          const started = new Set<number>()
-          let parentRuns = 0
-          let activityRuns = 0
-          let released = 0
-          const parentLayer = Parent.toLayer(() =>
-            Effect.suspend(() => {
-              parentRuns++
-              return Activity.make({
-                name: "fan-out",
-                success: Schema.Array(Schema.Number),
-                execute: Effect.suspend(() => {
-                  activityRuns++
-                  return Effect.forEach(
-                    Array.from({ length: childCount }, (_, index) => index),
-                    (index) => Child.execute({ index }),
-                    { concurrency }
-                  )
-                }).pipe(Effect.ensuring(Effect.sync(() => {
-                  released++
-                })))
-              })
-            })
-          )
-          const childLayer = Child.toLayer(({ index }) =>
-            Effect.gen(function*() {
-              started.add(index)
-              yield* DurableClock.sleep({ name: "child", duration: "2 seconds", inMemoryThreshold: Duration.zero })
-              return index
-            })
-          )
-          yield* Effect.gen(function*() {
-            const executionId = yield* Parent.execute(undefined, { discard: true })
-            for (const expected of waves) {
-              yield* awaitResult(Parent, executionId, "Suspended", () => started.size === expected)
-              assert.deepStrictEqual([...started].sort(), Array.from({ length: expected }, (_, index) => index))
-              assert.strictEqual(released, activityRuns)
-              assert.isAtMost(activityRuns, parentRuns)
-              if (expected === waves[0]) assert.strictEqual(parentRuns, 1)
-              yield* TestClock.adjust("2 seconds")
-              yield* settle
-            }
-            const result = yield* awaitResult(Parent, executionId, "Complete")
-            assert.deepStrictEqual(
-              result,
-              new Workflow.Complete({
-                exit: Exit.succeed(Array.from({ length: childCount }, (_, index) => index))
-              })
-            )
-            assert.isAtLeast(parentRuns, waves.length + 1)
-          }).pipe(Effect.provide(Layer.mergeAll(parentLayer, childLayer).pipe(Layer.provideMerge(engineLayer))))
-        }))
-    }
-
-    it.effect("child completions during activity cleanup wake the parent", () =>
-      Effect.gen(function*() {
-        const release = yield* Effect.makeLatch()
-        let cleaningUp = false
-        const gate = DurableDeferred.make("cleanup-gate")
-        const Parent = Workflow.make({
-          name: "BackportCleanupParent",
-          payload: {},
-          success: Schema.Array(Schema.Number),
-          idempotencyKey: () => "parent"
-        })
-        const Child = Workflow.make({
-          name: "BackportCleanupChild",
-          payload: { index: Schema.Number },
-          success: Schema.Number,
-          idempotencyKey: ({ index }) => String(index)
-        })
-        const parentLayer = Parent.toLayer(() =>
-          Activity.make({
-            name: "children",
-            success: Schema.Array(Schema.Number),
-            execute: Effect.forEach([0, 1], (index) => Child.execute({ index }), { concurrency: "unbounded" }).pipe(
-              Effect.ensuring(
-                Effect.sync(() => {
-                  cleaningUp = true
-                }).pipe(Effect.zipRight(release.await))
-              )
-            )
-          })
-        )
-        const childLayer = Child.toLayer(({ index }) => DurableDeferred.await(gate).pipe(Effect.as(index)))
-        yield* Effect.gen(function*() {
-          const executionId = yield* Parent.execute(undefined, { discard: true })
-          for (let i = 0; i < 100 && !cleaningUp; i++) yield* settle
-          assert.isTrue(cleaningUp)
-          for (const index of [0, 1]) {
-            const childId = yield* Child.executionId({ index })
-            yield* DurableDeferred.succeed(gate, {
-              token: DurableDeferred.tokenFromExecutionId(gate, { workflow: Child, executionId: childId }),
-              value: undefined
-            })
-            yield* awaitResult(Child, childId, "Complete")
-          }
-          yield* release.open
-          assert.deepStrictEqual(
-            yield* awaitResult(Parent, executionId, "Complete"),
-            new Workflow.Complete({ exit: Exit.succeed([0, 1]) })
-          )
-        }).pipe(
-          Effect.ensuring(release.open),
-          Effect.provide(Layer.mergeAll(parentLayer, childLayer).pipe(Layer.provideMerge(engineLayer)))
-        )
-      }))
-
-    it.effect("activity suspension survives beyond the ordinary interruption retry budget", () =>
-      Effect.gen(function*() {
-        const gate = DurableDeferred.make("retry-gate")
-        let attempts = 0
-        const Suspends = Workflow.make({ name: "BackportSuspensionBudget", payload: {}, idempotencyKey: () => "one" })
-        const layer = Suspends.toLayer(() =>
-          Activity.make({
-            name: "suspend",
-            execute: Effect.suspend(() => {
-              attempts++
-              return DurableDeferred.await(gate)
-            })
-          })
-        ).pipe(Layer.provideMerge(engineLayer))
-        yield* Effect.gen(function*() {
-          const executionId = yield* Suspends.execute(undefined, { discard: true })
-          yield* settle
-          yield* TestClock.adjust("2 minutes")
-          const result = yield* awaitResult(Suspends, executionId, "Suspended")
-          assert.strictEqual(result?._tag, "Suspended")
-          assert.strictEqual(attempts, 1)
-          yield* DurableDeferred.succeed(gate, {
-            token: DurableDeferred.tokenFromExecutionId(gate, { workflow: Suspends, executionId }),
-            value: undefined
-          })
-          assert.deepStrictEqual(
-            yield* awaitResult(Suspends, executionId, "Complete"),
-            new Workflow.Complete({ exit: Exit.void })
-          )
-        }).pipe(Effect.provide(layer))
-      }))
-
-    for (const mode of ["mapped", "into", "multi-await", "active-wins"] as const) {
-      it.effect(`deferred race ${mode} preserves the winning branch and replay`, () =>
-        Effect.gen(function*() {
-          const gate = DurableDeferred.make("race-gate", { success: Schema.String })
-          const second = DurableDeferred.make("second-gate", { success: Schema.String })
-          const auxiliary = DurableDeferred.make("auxiliary", { success: Schema.String })
-          const readGate = yield* Effect.makeLatch()
-          let readSecond = false
-          const active = yield* Effect.makeLatch()
-          let runs = 0
-          let activeCompleted = false
-          const Race = Workflow.make({
-            name: "BackportRace",
-            payload: {},
-            success: Schema.String,
-            idempotencyKey: () => "race"
-          })
-          const layer = Race.toLayer(() =>
-            Effect.gen(function*() {
-              runs++
-              const engine = yield* WorkflowEngineModule.WorkflowEngine
-              const branch = mode === "multi-await"
-                ? Effect.gen(function*() {
-                  const first = yield* DurableDeferred.await(gate)
-                  const next = yield* DurableDeferred.await(second)
-                  return `${first}:${next}`
-                })
-                : DurableDeferred.await(gate).pipe(Effect.map((value) => `mapped:${value}`))
-              return yield* DurableDeferred.raceAll({
-                name: "race",
-                success: Schema.String,
-                error: Schema.Never,
-                effects: [
-                  mode === "into" ? DurableDeferred.into(branch, auxiliary) : branch,
-                  Activity.make({
-                    name: "slow",
-                    success: Schema.String,
-                    execute: Effect.gen(function*() {
-                      yield* active.open
-                      yield* Effect.sleep("30 seconds")
-                      activeCompleted = true
-                      return "active"
-                    })
-                  })
-                ]
-              }).pipe(Effect.provideService(WorkflowEngineModule.WorkflowEngine, {
-                ...engine,
-                deferredResult: (deferred) =>
-                  engine.deferredResult(deferred).pipe(Effect.tap(() =>
-                    deferred.name === gate.name ? readGate.open : Effect.sync(() => {
-                      if (deferred.name === second.name) readSecond = true
-                    })
-                  ))
-              }))
-            })
-          ).pipe(Layer.provideMerge(engineLayer))
-          yield* Effect.gen(function*() {
-            const executionId = yield* Race.execute(undefined, { discard: true })
-            yield* active.await
-            yield* readGate.await
-            yield* settle
-            if (mode === "active-wins") {
-              yield* TestClock.adjust("30 seconds")
-            } else {
-              yield* DurableDeferred.succeed(gate, {
-                token: DurableDeferred.tokenFromExecutionId(gate, { workflow: Race, executionId }),
-                value: "signal"
-              })
-              if (mode === "multi-await") {
-                for (let i = 0; i < 100 && !readSecond; i++) yield* settle
-                assert.isTrue(readSecond)
-                yield* DurableDeferred.succeed(second, {
-                  token: DurableDeferred.tokenFromExecutionId(second, { workflow: Race, executionId }),
-                  value: "second"
-                })
-              }
-            }
-            const result = yield* awaitResult(Race, executionId, "Complete")
-            const expected = mode === "active-wins"
-              ? "active"
-              : mode === "multi-await"
-              ? "signal:second"
-              : "mapped:signal"
-            assert.deepStrictEqual(result, new Workflow.Complete({ exit: Exit.succeed(expected) }))
-            assert.strictEqual(activeCompleted, mode === "active-wins")
-            if (mode !== "active-wins") assert.isAtLeast(runs, mode === "multi-await" ? 3 : 2)
-            assert.strictEqual(yield* Race.execute(undefined), expected)
-          }).pipe(Effect.provide(layer))
-        }))
-    }
-
-    it.effect("DurableDeferred.into does not persist an interrupted attempt", () =>
-      Effect.gen(function*() {
-        const deferred = DurableDeferred.make("interrupted", { success: Schema.String })
-        const instance = WorkflowEngineModule.WorkflowInstance.initial(
-          Workflow.make({ name: "BackportInterrupted", payload: {}, idempotencyKey: () => "one" }),
-          "interrupted"
-        )
-        const engine = yield* WorkflowEngineModule.WorkflowEngine
-        const writes: Array<Exit.Exit<unknown, unknown>> = []
-        const exit = yield* DurableDeferred.into(Effect.interrupt, deferred).pipe(
-          Effect.provideService(WorkflowEngineModule.WorkflowInstance, instance),
-          Effect.provideService(WorkflowEngineModule.WorkflowEngine, {
-            ...engine,
-            deferredDone: (_deferred, options) =>
-              Effect.sync(() => {
-                writes.push(options.exit)
-              })
-          }),
-          Effect.exit
-        )
-        assert.isTrue(Exit.isInterrupted(exit))
-        assert.deepStrictEqual(writes, [])
-      }).pipe(Effect.provide(engineLayer)))
-  })
 })
