+++ packages/cluster/test/Interruptors.test.ts
@@ -1,7 +1,7 @@
 import { EntityAddress, EntityId, EntityType, ShardId } from "@effect/cluster"
 import { assert, describe, it } from "@effect/vitest"
 import { Effect, Exit } from "effect"
-import * as Teardown from "../src/internal/interruptors.js"
+import { make } from "../src/internal/interruptors.js"
 
 describe("interruptors", () => {
   const address = EntityAddress.make({
@@ -11,28 +11,25 @@ describe("interruptors", () => {
   })
   it.effect("nested teardown owners release only their own reference", () =>
     Effect.gen(function*() {
+      const Teardown = make()
       Teardown.acquireEntity(address)
       Teardown.acquireEntity(address)
-      try {
-        Teardown.releaseEntity(address)
-        assert.isTrue(Teardown.isActive(address))
-        yield* Teardown.aroundEntityType(
-          address.entityType,
-          Effect.gen(function*() {
-            Teardown.releaseEntity(address)
-            assert.isTrue(Teardown.isActive(address))
-            yield* Teardown.aroundShard(address.shardId, Effect.void)
-            assert.isTrue(Teardown.isActive(address))
-          })
-        )
-        assert.isFalse(Teardown.isActive(address))
-      } finally {
-        Teardown.releaseEntity(address)
-        Teardown.releaseEntity(address)
-      }
+      Teardown.releaseEntity(address)
+      assert.isTrue(Teardown.isActive(address))
+      yield* Teardown.aroundEntityType(
+        address.entityType,
+        Effect.gen(function*() {
+          Teardown.releaseEntity(address)
+          assert.isTrue(Teardown.isActive(address))
+          yield* Teardown.aroundShard(address.shardId, Effect.void)
+          assert.isTrue(Teardown.isActive(address))
+        })
+      )
+      assert.isFalse(Teardown.isActive(address))
     }))
   it.effect("interrupted teardown releases the active classification", () =>
     Effect.gen(function*() {
+      const Teardown = make()
       const exit = yield* Teardown.aroundShard(
         address.shardId,
         Effect.gen(function*() {
