+++ packages/workflow/tsconfig.test.json
@@ -5,6 +5,6 @@
   "compilerOptions": {
     "tsBuildInfoFile": ".tsbuildinfo/test.tsbuildinfo",
     "rootDir": "test",
-    "noEmit": true
+    "outDir": "build/test"
   }
 }
